refactor(mcp): extract the two dialogs out of mcp-server-list
mcp-server-list.tsx concentrated all the MCP-connection churn in ~985 lines
(#569, related #461). Its two dialogs - the connect/edit draft and the
tool picker - move into their own components, and the list drops to ~765
lines. A pure move: the JSX is unchanged, the parent keeps every useState and
both handlers (passed as onSubmit / onSave), and no copy changes.

The shapes the three files share - Scope, DraftAuth, DraftState,
ToolPickerState and the SCOPE_LABEL key map - move into a leaf module
mcp-server-list-types.ts, so the dialogs read them without importing the
component that renders them. Importing SCOPE_LABEL (a value) back from the
list would have run a dialog module while the list module was still
initializing; a leaf module has no such cycle.

AUTH_CHOICES and its hint lookup were used only by the draft dialog, so they
move with it. The i18n items the audit also listed for this file were already
migrated (#425/#446): check:i18n is clean, nothing left to translate.

17 integration tests still green; tsc, eslint, prettier, i18n all clean.

Part-of #569
